+++ src/TelegramBotChannel.php
@@ -3,6 +3,7 @@
 namespace AlexItDev91\LaravelTelegramBot;
 
 use Override;
+use AlexItDev91\LaravelTelegramBot\Concerns\SendsTelegramMessageShortcuts;
 use AlexItDev91\LaravelTelegramBot\Contracts\TelegramBotClient;
 use AlexItDev91\LaravelTelegramBot\DTO\TelegramBotMethodRequest;
 use AlexItDev91\LaravelTelegramBot\DTO\TelegramBotRequestData;
@@ -16,6 +17,7 @@ class TelegramBotChannel implements TelegramBotClient
 {
     use TelegramBotApiMethods;
     use TelegramBotTypedApiMethods;
+    use SendsTelegramMessageShortcuts;
 
     public function __construct(
         private readonly TelegramBotClient $bot,